Refactor checkPuzzle and checkPuzzle tests
Refactored checkPuzzle to take cells and solution arguments rather than row and column hints. The function still works by comparing the hints though otherwise a correct solution for the given hints could return false when the puzzle's solution is not unique.

If the generated puzzles could be guaranteed to be unique, checkPuzzle could be simplified to just compare the current cells and the solution cells